Added support for nodejs backend within project, updated file structure and removed unused assets
