+++ client/public/assets/images/bridges/bridges-bottom-left.png
