+++ client/public/assets/images/bridges/bridges-bottom-right.png
