+++ client/public/assets/images/bridges/bridges-top-left.png
