+++ client/public/assets/images/hexes/brick.png
