+++ client/public/assets/images/hexes/desert.png
