+++ client/public/assets/images/hexes/ore.png
