+++ client/public/assets/images/hexes/sheep.png
