+++ client/public/assets/images/hexes/wood.png
