+++ client/public/assets/images/pirate-ship.png
