+++ client/src/index.css
@@ -93,7 +93,7 @@ code {
   position: absolute;
   width: 80px;
   height: 108px;
-  background-image: url('../public/assets/pirate-ship.png');
+  background-image: url('../public/assets/images/pirate-ship.png');
   background-size: cover;
   transition: all 0.3s ease; /* Updated for smooth transition */
 }
