+++ client/src/modes/base/game.js
@@ -2,9 +2,9 @@ import React, { useState, useEffect } from "react";
 import { HexGrid, Layout, Pattern, Hexagon } from "react-hexgrid";
 import NumberTileBackground from '../../components/NumberTileBackground';
 import Tile from '../../components/Tile';
-import { Hexes, Bridges, Ports } from './constants';
+import { Hexes, Bridges, Ports } from './constants'
 
-const FourIslands = (props) => {
+const BaseGame = (props) => {
   const [boardLayout, setBoardLayout] = useState(Hexes);
   const [bridges,] = useState(Bridges)
   const [ports,] = useState(Ports)
@@ -17,7 +17,9 @@ const FourIslands = (props) => {
   useEffect(() => {
     function handleKeyDown(e) {
       e.preventDefault();
-      if (e.keyCode === 51) {
+
+      // Key: 1
+      if (e.keyCode === 49) {
         shuffleBoard();
       }
     }
@@ -59,17 +61,17 @@ const FourIslands = (props) => {
     }
 
     return (
-        pipCounts.forest >= 10 &&
-        pipCounts.forest <= 15 &&
-        pipCounts.brick >= 7 &&
-        pipCounts.brick <= 12 &&
-        pipCounts.sheep >= 10 &&
-        pipCounts.sheep <= 15 &&
-        pipCounts.wheat >= 10 &&
-        pipCounts.wheat <= 15 &&
-        pipCounts.ore >= 7 &&
-        pipCounts.ore <= 12
-      );
+      pipCounts.forest >= 12 &&
+      pipCounts.forest <= 15 &&
+      pipCounts.brick >= 9 &&
+      pipCounts.brick <= 12 &&
+      pipCounts.sheep >= 12 &&
+      pipCounts.sheep <= 15 &&
+      pipCounts.wheat >= 12 &&
+      pipCounts.wheat <= 15 &&
+      pipCounts.ore >= 9 &&
+      pipCounts.ore <= 12
+    );
   }
 
   function shuffleArray(array) {
@@ -90,33 +92,29 @@ const FourIslands = (props) => {
   function placeSixesAndEights() {
     let sixesAndEights = [6, 6, 8, 8];
     shuffleArray(sixesAndEights);
-  
-    // Initialize a count for 6s and 8s for each resource type
-    const resourceCounts = { forest: 0, brick: 0, sheep: 0, wheat: 0, ore: 0 };
-  
-    // Helper function to check if a resource type can receive another 6 or 8
-    const canPlaceNumber = (resource, number) => {
-      const counts = Object.values(resourceCounts);
-      const minCount = Math.min(...counts);
-      return resourceCounts[resource] === minCount || (number === 8 && resourceCounts[resource] === minCount + 1);
-    };
-  
+
+    let resourceSixOrEight = { forest: false, brick: false, sheep: false, wheat: false, ore: false };
+
+    // Create an array of indices for hexagons that are not desert
+    let nonDesertIndices = boardLayout
+      .map((hex, index) => hex.fill !== "desert" ? index : -1)
+      .filter(index => index !== -1);
+
     for (let number of sixesAndEights) {
-      let nonDesertHexes = boardLayout.filter(hex => hex.fill !== 'desert' && hex.number === null && !isNeighborWithSixOrEight(hex.id));
-      shuffleArray(nonDesertHexes);
-  
+      shuffleArray(nonDesertIndices);
       let placed = false;
-      for (let hex of nonDesertHexes) {
-        if (canPlaceNumber(hex.fill, number)) {
+
+      for (let index of nonDesertIndices) {
+        let hex = boardLayout[index];
+        if (hex.number === null && !isNeighborWithSixOrEight(hex.id) && !resourceSixOrEight[hex.fill]) {
           hex.number = number;
-          resourceCounts[hex.fill]++;
           placed = true;
+          resourceSixOrEight[hex.fill] = true; // Mark this resource as having a 6 or 8
           break;
         }
       }
-  
+
       if (!placed) {
-        // Reset and retry if unable to place number with proper distribution
         boardLayout.forEach(hex => {
           if (hex.fill !== "desert") hex.number = null;
         });
@@ -125,7 +123,6 @@ const FourIslands = (props) => {
       }
     }
   }
-  
 
   function shuffleFills() {
     let validFills = false;
@@ -134,13 +131,7 @@ const FourIslands = (props) => {
     while (!validFills && attempts < 1000) {
       // Shuffle 'fill' attributes including the desert
       const fills = boardLayout.map(hex => hex.fill);
-      for (let i = fills.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1));
-        if (fills[i] === 'ocean' || fills[j] === 'ocean') {
-            continue;
-        }
-        [fills[i], fills[j]] = [fills[j], fills[i]];
-      }
+      shuffleArray(fills);
       for (let i = 0; i < boardLayout.length; i++) {
         boardLayout[i].fill = fills[i];
       }
@@ -157,16 +148,17 @@ const FourIslands = (props) => {
   }
 
   function shufflePorts() {
-    const fills = ports.map(port => port.fill);
-    shuffleArray(fills);
-    for (let i = 0; i < ports.length; i++) {
-      ports[i].fill = fills[i];
-    }
+      // Shuffle 'fill' attributes including the desert
+      const fills = ports.map(port => port.fill);
+      shuffleArray(fills);
+      for (let i = 0; i < ports.length; i++) {
+        ports[i].fill = fills[i];
+      }
   }
 
   function checkFillValidity() {
     for (const hex of boardLayout) {
-      if (hex.fill === "desert") continue;
+      if (hex.fill === "desert" || hex.fill === "any-top-left") continue;
       let sameTypeCount = 0;
 
       for (const neighborId of hex.neighbors) {
@@ -189,11 +181,11 @@ const FourIslands = (props) => {
   }
 
   function fillInOtherNumbers() {
-    let otherNumbers = [2, 3, 3, 4, 4, 4, 5, 5, 9, 9, 9, 10, 10, 10, 11, 11, 12];
+    let otherNumbers = [2, 3, 3, 4, 4, 5, 5, 9, 9, 10, 10, 11, 11, 12];
     shuffleArray(otherNumbers);
 
     for (const hex of boardLayout) {
-      if (hex.fill === "desert" || hex.fill === "ocean" || hex.number !== null) continue;
+      if (hex.fill === "desert" || hex.number !== null) continue;
 
       let placed = false;
       for (let i = 0; i < otherNumbers.length; i++) {
@@ -224,7 +216,7 @@ const FourIslands = (props) => {
       attempts++;
       // Reset numbers
       boardLayout.forEach(
-        (hex) => (hex.number = (hex.fill === "desert") ? null : undefined)
+        (hex) => (hex.number = (hex.fill === "desert" || hex.fill === "any-top-left") ? null : undefined)
       );
 
       placeSixesAndEights();
@@ -234,20 +226,11 @@ const FourIslands = (props) => {
       if (success) {
         success = isValidPipDistribution();
       }
-    } while (!success && attempts < 500);
+    } while (!success && attempts < 100);
 
-    if (attempts >= 500 || !success) {
+    if (attempts >= 100 || !success) {
       console.log(`Failed to find a valid distribution after ${attempts} attempts`);
     }
-  }
-
-  function shufflePorts() {
-      // Shuffle 'fill' attributes including the desert
-      const fills = ports.map(port => port.fill);
-      shuffleArray(fills);
-      for (let i = 0; i < ports.length; i++) {
-        ports[i].fill = fills[i];
-      }
   }
 
   function shuffleBoard() {
@@ -263,7 +246,7 @@ const FourIslands = (props) => {
       <div style={{ display: "flex", alignItems: "center", justifyContent: "center", height: "100vh" }}>
         <div className="hexgrid-container">
           <div className="board" style={{zIndex: -1, position: "absolute"}}>
-            <HexGrid width={1000} height={1000} viewBox="-70 -70 140 140"
+            <HexGrid width={1000} height={1000} viewBox="-60 -60 120 120"
               style={{ transform: "rotate(90deg)" }}>
               <Layout
                 size={{ x: 10, y: 10 }}
@@ -274,8 +257,8 @@ const FourIslands = (props) => {
                 {boardLayout.map((hex, index) => (
                   <Hexagon
                     key={index}
-                    q={hex.q - 1}
-                    r={hex.r + 0.5}
+                    q={hex.q}
+                    r={hex.r}
                     s={hex.s}
                     stroke="black"
                     strokeWidth={0.2}
@@ -290,19 +273,17 @@ const FourIslands = (props) => {
                     )}
                   </Hexagon>
                 ))}
-                <Pattern id="forest" link="/assets/wood.png" />
-                <Pattern id="ore" link="/assets/ore.png" />
-                <Pattern id="wheat" link="/assets/wheat.png" />
-                <Pattern id="sheep" link="/assets/sheep.png" />
-                <Pattern id="brick" link="/assets/brick.png" />
-                <Pattern id="desert" link="/assets/desert.png" />
-                <Pattern id="ocean" link="/assets/ocean.png" />
+                <Pattern id="forest" link="/assets/images/hexes/wood.png" />
+                <Pattern id="ore" link="/assets/images/hexes/ore.png" />
+                <Pattern id="wheat" link="/assets/images/hexes/wheat.png" />
+                <Pattern id="sheep" link="/assets/images/hexes/sheep.png" />
+                <Pattern id="brick" link="/assets/images/hexes/brick.png" />
+                <Pattern id="desert" link="/assets/images/hexes/desert.png" />
               </Layout>
             </HexGrid>
           </div>
-        </div>
-        <div className="bridges" style={{position: "absolute", zIndex: 1}}>
-          <HexGrid width={1000} height={1000} viewBox="-70 -70 140 140"
+          <div className="bridges" style={{position: "absolute", zIndex: 1}}>
+          <HexGrid width={1000} height={1000} viewBox="-60 -60 120 120"
               style={{ transform: "rotate(90deg)" }}>
               <Layout
                 size={{ x: 10, y: 10 }}
@@ -313,24 +294,24 @@ const FourIslands = (props) => {
                 {bridges.map((bridge, index) => (
                   <Hexagon
                     key={index}
-                    q={bridge.q - 1}
-                    r={bridge.r + 0.5}
+                    q={bridge.q}
+                    r={bridge.r}
                     s={bridge.s}
                     fill={bridge.fill}
                   >
                   </Hexagon>
                 ))}
-                <Pattern id="bridges-top-left" link="/assets/bridges-bottom-left.png" />
-                <Pattern id="bridges-top-right" link="/assets/bridges-top-left.png" />
-                <Pattern id="bridges-left" link="/assets/bridges-bottom.png" />
-                <Pattern id="bridges-right" link="/assets/bridges-top.png" />
-                <Pattern id="bridges-bottom-left" link="/assets/bridges-bottom-right.png" />
-                <Pattern id="bridges-bottom-right" link="/assets/bridges-top-right.png" />
+                <Pattern id="bridges-top-left" link="/assets/images/bridges/bridges-bottom-left.png" />
+                <Pattern id="bridges-top-right" link="/assets/images/bridges/bridges-top-left.png" />
+                <Pattern id="bridges-left" link="/assets/images/bridges/bridges-bottom.png" />
+                <Pattern id="bridges-right" link="/assets/images/bridges/bridges-top.png" />
+                <Pattern id="bridges-bottom-left" link="/assets/images/bridges/bridges-bottom-right.png" />
+                <Pattern id="bridges-bottom-right" link="/assets/images/bridges/bridges-top-right.png" />
               </Layout>
             </HexGrid>
           </div>
           <div className="ports" style={{position: "absolute", zIndex: 1}}>
-            <HexGrid width={1000} height={1000} viewBox="-70 -70 140 140"
+            <HexGrid width={1000} height={1000} viewBox="-60 -60 120 120"
               style={{ transform: "rotate(90deg)" }}>
               <Layout
                 size={{ x: 10, y: 10 }}
@@ -341,25 +322,26 @@ const FourIslands = (props) => {
                 {ports.map((port, index) => (
                   <Hexagon
                     key={index}
-                    q={port.q - 1}
-                    r={port.r + 0.5}
+                    q={port.q}
+                    r={port.r}
                     s={port.s}
                     fill={port.fill}
                   >
                   </Hexagon>
                 ))}
-                <Pattern id="any" link="/assets/any.png" />
-                <Pattern id="ore-port" link="/assets/ore-port.png" />
-                <Pattern id="wheat-port" link="/assets/wheat-port.png" />
-                <Pattern id="sheep-port" link="/assets/sheep-port.png" />
-                <Pattern id="brick-port" link="/assets/brick-port.png" />
-                <Pattern id="wood-port" link="/assets/wood-port.png" />
+                <Pattern id="any" link="/assets/images/ports/any-port.png" />
+                <Pattern id="ore-port" link="/assets/images/ports/ore-port.png" />
+                <Pattern id="wheat-port" link="/assets/images/ports/wheat-port.png" />
+                <Pattern id="sheep-port" link="/assets/images/ports/sheep-port.png" />
+                <Pattern id="brick-port" link="/assets/images/ports/brick-port.png" />
+                <Pattern id="wood-port" link="/assets/images/ports/wood-port.png" />
               </Layout>
             </HexGrid>
           </div>
+        </div>
       </div>
     </>
   );
 };
 
-export default FourIslands;
+export default BaseGame;
