+++ server/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "catan-board-generator_server",
+  "version": "0.1.0",
+  "description": "Settlers of Catan Project",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.js"
+  },
+  "author": "Jake Krummenacher and Sam McGrath",
+  "license": "MIT",
+  "dependencies": {
+    "express": "^4.18.2",
+    "socket.io": "^4.7.4"
+  }
+}
